+++ src/redux/store.js
@@ -1,0 +1,11 @@
+import { configureStore } from '@reduxjs/toolkit';
+import eventsReducer from './events';
+//store config and return
+const store = configureStore({
+  reducer: {
+    events: eventsReducer,
+  },
+  devTools: true,
+});
+
+export default store;
